requirements.txt
 Flask 
werkzeug.utils 
 cv2
 ultralytics  
 numpy 
 os
 torch
 tensorflow
 time